% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ascn}
\alias{ascn}
\title{Example allele specific copy number (ASCN) data}
\format{
A \code{data.frame}
}
\source{
Generate from \verb{data-raw/}, raw source from our study by
calling ASCAT v3.0 alpha on corresponding WES sequencing data.
}
\description{
Example allele specific copy number (ASCN) data
}
\examples{
data("ascn")
}
