% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clusterGPosition}
\alias{clusterGPosition}
\title{Cluster genomic regions by distance from region center}
\usage{
clusterGPosition(dt, distance = 10000000)
}
\arguments{
\item{dt}{a genomic region data with first 3 columns for
chromosome, start and end.}

\item{distance}{a distance value for clustering,
regions with average distance lower than this value would be clusetered
into one. Should not greater than 1e8.}
}
\value{
cluster list.
}
\description{
Cluster genomic regions by distance from region center
}
\examples{
library(data.table)
region <- data.table(
  chr = c("chr1", "chr2", "chr1"),
  start = c(1032, 992, 10000), end = c(4242, 1e6, 400023)
)
clusterGPosition(region, distance = 1000)
clusterGPosition(region)
}
