% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertID.R
\name{convertID}
\alias{convertID}
\title{Convert Gene IDs between Ensembl and Hugo Symbol System}
\usage{
convertID(
  IDs,
  type = c("ensembl", "symbol"),
  genome_build = c("hg38", "hg19", "mm10", "mm9"),
  multiple = FALSE
)
}
\arguments{
\item{IDs}{a character vector to convert.}

\item{type}{type of input \code{IDs}, could be 'ensembl' or 'symbol'.}

\item{genome_build}{reference genome build.}

\item{multiple}{if \code{TRUE}, return a \code{data.table} instead of a string vector,
so multiple identifier mappings can be kept.}
}
\value{
a vector or a \code{data.table}.
}
\description{
Convert Gene IDs between Ensembl and Hugo Symbol System
}
\examples{
\donttest{
convertID("ENSG00000243485")
convertID("ENSG00000243485", multiple = TRUE)
convertID(c("TP53", "KRAS", "EGFR", "MYC"), type = "symbol")
}
}
