% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy}
\alias{deploy}
\title{Deploy Command Line Interface to System Local Path}
\usage{
deploy(target = "/usr/local/bin")
}
\arguments{
\item{target}{the target path to deploy the CLI.}
}
\value{
Nothing.
}
\description{
Only should be used in Unix-like system.
For details of the arguments passing to CLI, please check \code{\link[=gcap.workflow]{gcap.workflow()}}
and \code{\link[=gcap.ASCNworkflow]{gcap.ASCNworkflow()}}.
}
