% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{fselect}
\alias{fselect}
\title{Auto-select features}
\usage{
fselect(data, target, positive, seed = 2021L)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{target}{target column name}

\item{positive}{positive case representation, like "1", "+", etc.}

\item{seed}{random seed}
}
\value{
a \code{list}
}
\description{
Auto-select features
}
