% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcap-package.R
\docType{package}
\name{gcap-package}
\alias{gcap}
\alias{gcap-package}
\title{gcap: Gene-level Circular Amplicon Prediction}
\description{
Provides data processing pipeline feeding paired bam files
    and XGBOOST model for predicting tumor circular amplicons (also known
    as ecDNA) in gene level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShixiangWang/gcap}
  \item Report bugs at \url{https://github.com/ShixiangWang/gcap/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
