% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring2.R
\name{gcap.aggrProb}
\alias{gcap.aggrProb}
\title{Determine sample-level ecDNA status with aggregated probability}
\usage{
gcap.aggrProb(
  data,
  genome_build = c("hg38", "hg19"),
  min_ngene = 2L,
  min_ncluster = 2L
)
}
\arguments{
\item{data}{a \code{data.table} containing result from \link{gcap.runPrediction}.
The column storing prediction result must start with \code{pred}.}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}

\item{min_ngene, min_ncluster}{minimal gene or cluster number hosted on
amplicon. In logic, at least \code{2}.}
}
\value{
a \code{data.table}.
\itemize{
\item \verb{*_gene_based_prob} for gene based probability.
\item \verb{*_cluster_based_prob} for cluster based probability.
We cluster genes based on  distance (hard cutoff \code{1e6} (i.e., 1Mb)).
For genes from different
chromosomes, the distance set to \code{1e8} (i.e., 100Mb).
}
}
\description{
Aggregated probability is calculated from gene prob or cluster prob
(this first cluster genes and then aggregate the cluster prob).
}
\examples{
data("ec")
ec2 <- ec
ec2$pred <- gcap.runPrediction(ec)
score <- gcap.aggrProb(ec2)
score
}
