% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCNworkflow.R
\name{gcap.ASCNworkflow}
\alias{gcap.ASCNworkflow}
\title{GCAP workflow for gene-level amplicon prediction from ASCN input}
\usage{
gcap.ASCNworkflow(
  data,
  genome_build = c("hg38", "hg19"),
  target = c("circle", "nonLinear"),
  prob_cutoff = 0.9,
  outdir = getwd(),
  result_file_prefix = paste0("gcap_", uuid::UUIDgenerate(TRUE))
)
}
\arguments{
\item{data}{a \code{data.frame} with following rows. The key columns can be obtained
from common allele specific CNV calling software, e.g., ASCAT, Sequenza, FACETS.
\itemize{
\item chromosome: chromosome names starts with 'chr'.
\item start: start position of the segment.
\item end: end position of the segment.
\item total_cn: total integer copy number of the segment.
\item minor_cn: minor allele integer copy number of the segment.
\item sample: sample identifier.
\item purity: tumor purity of the sample. Set to \code{1} if you don't know.
\item ploidy (optinal): ploidy value of the sample tumor genome.
\item age (optional): age of the case, use along with \code{gender}.
\item gender (optional): gender of the case, use along with \code{age}.
}}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}

\item{target}{'circle' and 'nonLinear' to select model if predict circle
amplicon or non-linear amplicon. \strong{Can be one or both}.}

\item{prob_cutoff}{a value in \verb{[0-1]} to determine the cutoff for labeling
response. The default cutoff \code{0.9} is a balance between F1 score and amplicon
load accuracy.}

\item{outdir}{result output path.}

\item{result_file_prefix}{file name prefix (without directory path) for storing
final model prediction file in CSV format.
Default a unique file name is generated by UUID approach.}
}
\value{
a list of invisible \code{data.table} and corresponding files saved to local machine.
}
\description{
Unlike \link{gcap.workflow}, this function directly uses the allele-specific
copy number data along with some extra sample information to infer
ecDNA genes.
}
\examples{
data("ascn")
data <- ascn
rv <- gcap.ASCNworkflow(data, outdir = tempdir())
data$purity <- 1
rv2 <- gcap.ASCNworkflow(data, outdir = tempdir())
data$age <- 60
data$gender <- "XY"
rv3 <- gcap.ASCNworkflow(data, outdir = tempdir())
}
