% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse2Genes.R
\name{gcap.collapse2Genes}
\alias{gcap.collapse2Genes}
\title{Generate unified gene-level feature data}
\usage{
gcap.collapse2Genes(
  fts,
  extra_info = NULL,
  include_type = FALSE,
  fix_type = TRUE,
  genome_build = c("hg38", "hg19", "mm10"),
  overlap = 1
)
}
\arguments{
\item{fts}{(modified) result from \code{\link[=gcap.extractFeatures]{gcap.extractFeatures()}}}

\item{extra_info}{(optional) a \code{data.frame} with 3 columns 'sample',
'age' and 'gender', for including cancer type, check
parameter \code{include_type}. For gender, should be 'XX' or 'XY',
also could be \code{0} for 'XX' and \code{1} for 'XY'.}

\item{include_type}{if \code{TRUE}, a fourth column named 'type'
should be included in \code{extra_info}, the supported cancer
type should be described with \href{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}{TCGA cancer type abbr.}.}

\item{fix_type}{default is \code{TRUE}, only cancer types used in pre-trained
models are used, others will be convert to \code{NA}. If \code{FALSE}, only generating
one-hot encoding for cancer types in input data.}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}

\item{overlap}{the overlap percentage on gene.}
}
\value{
a \code{data.table}.
}
\description{
Generate unified gene-level feature data
}
