% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{gcap.dotplot}
\alias{gcap.dotplot}
\title{Plot CN Mean vs. Frequency Dotplot}
\usage{
gcap.dotplot(
  fCNA,
  by = c("gene_id", "band", "chr"),
  filter = cn > 50 | (N > 1 & cn > 20),
  include = c("circular", "possibly_circular"),
  ...
)
}
\arguments{
\item{fCNA}{a \link{fCNA} object.}

\item{by}{the level of focal amplicons.}

\item{filter}{a filter expression based on \code{cn} (CN mean) and \code{N} (frequency).}

\item{include}{amplicon type to include for plotting.}

\item{...}{other parameters passing to \code{ggrepel::geom_label_repel()}.}
}
\value{
a ggplot.
}
\description{
Plot CN Mean vs. Frequency Dotplot
}
