% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{gcap.enrich}
\alias{gcap.enrich}
\title{Enrich analysis}
\usage{
gcap.enrich(
  data,
  target = c(NA, "circular", "noncircular"),
  class_by = NULL,
  analysis_func = c("enricher", "fgsea"),
  gene_encode = c("ensembl", "symbol"),
  species = "Homo sapiens",
  category = "H",
  subcategory = "",
  genome_build = c("hg38", "hg19"),
  ...
)
}
\arguments{
\item{data}{a \link{fCNA} object, or a vector (list) of genes.
If the vector is named, the names must be genes and values must numeric.}

\item{target}{\code{NA} for general, and "circular" or "noncircular" for specific fCNA type.}

\item{class_by}{if a \link{fCNA} object provided, this specify a column name in \code{fCNA$sample_summary} to
classify genes into different sample groups for enrichment. NOTE: any group with <10 genes is ignored.}

\item{analysis_func}{analysis function to run.}

\item{gene_encode}{gene ID type from input.}

\item{species}{species, not to change.}

\item{category}{gene set main category name.}

\item{subcategory}{gene set sub-category name.}

\item{genome_build}{genome version.}

\item{...}{other parameters passing to \code{clusterProfiler::enricher()}
or \code{fgsea::fgsea}.}
}
\value{
depends on \code{data} and \code{analysis_func}.
}
\description{
Enrich analysis
}
