% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_components.R
\name{gcap.extractComponents}
\alias{gcap.extractComponents}
\title{Decompose distribution into components}
\usage{
gcap.extractComponents(
  dat,
  dist = "norm",
  seed = 2021L,
  model_selection = "BIC",
  threshold = 0.4,
  min_prior = 0.001,
  niter = 1000,
  nrep = 5,
  min_comp = 1,
  max_comp = 10
)
}
\arguments{
\item{dat}{a distribution represented as a numeric vector to decompose.}

\item{dist}{'pois' or 'norm' (default).}

\item{seed}{seed number.}

\item{model_selection}{model selection strategy, default is 'BIC'.
Details about custom setting please refer to \strong{flexmix} package.}

\item{threshold}{default is \code{0.5}. Sometimes, the result components
include adjacent distributions with similar mu
(two and more distribution are very close), we use this threshold
to obtain a more meaningful fit with less components.}

\item{min_prior}{the minimum relative size of components, default is 0.001.
Details about custom setting please refer to \strong{flexmix} package.}

\item{niter}{the maximum number of iterations.}

\item{nrep}{number of run times for each value of component,
keep only the solution with maximum likelihood.}

\item{min_comp}{minimal number of components to fit, default is 1.}

\item{max_comp}{maximal number of components to fit, default is 10.}
}
\description{
Decompose distribution into components
}
\examples{
set.seed(2021)
x <- c(rnorm(10, 0), rnorm(50, 1), rnorm(20, 4), rnorm(5, 10))
y <- gcap.extractComponents(x, max_comp = 5)
y
}
