% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{gcap.plotDistribution}
\alias{gcap.plotDistribution}
\title{Draw Gene Amplicon Distribution By Gene or User Specified Class}
\usage{
gcap.plotDistribution(
  fCNA,
  x = NULL,
  x_size = 8,
  by = c("gene", "sample"),
  merge_circular = TRUE,
  fill = TRUE,
  palette = c("#CCCCCC", "#0066CC", "#FFCCCC", "#CC0033"),
  ...
)
}
\arguments{
\item{fCNA}{a \link{fCNA} object.}

\item{x}{a list of gene or a column name in \code{fCNA$sample_summary}.}

\item{x_size}{font size of x axis text.}

\item{by}{"gene" (default) or "sample" mode. When "sample" is set, it uses the column specified
by \code{x} to grouping samples.}

\item{merge_circular}{if \code{TRUE}, merge 'possibly_circular' into 'circular' type.}

\item{fill}{if \code{TRUE}, show the percentage instead of count.}

\item{palette}{color palette.}

\item{...}{other parameters passing to \code{ggplot2::geom_bar}.}
}
\value{
a ggplot object.
}
\description{
See \code{\link[=gcap.plotProfile]{gcap.plotProfile()}} for examples.
}
\seealso{
\link{gcap.plotProfile} for plot landscape of fCNA, \link{fCNA} for building object.
}
