% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForest.R
\name{gcap.plotForest}
\alias{gcap.plotForest}
\title{Draw Forest Plot for fCNA+/- Comparison by Sample or Gene}
\usage{
gcap.plotForest(
  fCNA,
  response_data,
  f = c("coxph", "binomial", "gaussian", "Gamma", "inverse.gaussian", "poisson",
    "quasi", "quasibinomial", "quasipoisson"),
  x = "class",
  covars = NULL,
  y = if (is.data.frame(response_data)) {     colnames(response_data)[-1] } else {    
    response_data },
  merge_circular = TRUE,
  x_is_gene = FALSE,
  gene_focus = c("fCNA", "circular"),
  ending_time = NULL,
  parallel = FALSE,
  exp = NULL,
  ref_line = NULL,
  xlim = NULL,
  ...
)
}
\arguments{
\item{fCNA}{a \link{fCNA} object.}

\item{response_data}{a \code{data.frame} or specified column names in \code{fCNA$sample_summary} prepared
regression modeling, default is CoxPH model, you can use same data as \link{gcap.plotKMcurve}.}

\item{f}{a length-1 string specifying modeling function or family of \code{\link[=glm]{glm()}}, default is 'coxph'. Other options are members of GLM family, see \code{\link[stats:family]{stats::family()}}.
'binomial' is logistic, and 'gaussian' is linear.}

\item{x}{focal variables (terms), the column names of \code{fCNA$sample_summary} or a list of genes.}

\item{covars}{covariables in \code{fCNA$sample_summary}.}

\item{y}{predicted variables or expression in string format,
e.g., \code{"sex"} or \code{"factor(sex)"} is acceptable.}

\item{merge_circular}{if \code{TRUE}, merge 'possibly_circular' into 'circular' type.}

\item{x_is_gene}{if \code{TRUE}, treat \code{x} as a list of genes.}

\item{gene_focus}{focal amplication type you focus on.
Can be 'fCNA' or 'circular'. If 'fCNA' selected,
noncircular and circular genes are included to classify samples.}

\item{ending_time}{survival analysis ending time. If a numeric ending
is typed, all survival data longer than the ending time will be rewritten.}

\item{parallel}{if \code{TRUE}, use N-1 cores to run the task.}

\item{exp}{logical, indicating whether or not to exponentiate the the coefficients.}

\item{ref_line}{reference line, default is 1 for HR.}

\item{xlim}{limits of x axis.}

\item{...}{other plot options passing to \code{?forestploter::forest()}.
Also check \url{https://github.com/adayim/forestploter} to see more complex adjustment of the result plot.}
}
\value{
a list containing model and forest plot.
}
\description{
See \code{\link[=gcap.plotProfile]{gcap.plotProfile()}} for examples.
}
