% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKMcurve.R
\name{gcap.plotKMcurve}
\alias{gcap.plotKMcurve}
\title{Draw K-M Curve for fCNA Survival Comparison by Sample or Gene}
\usage{
gcap.plotKMcurve(
  fCNA,
  surv_data,
  merge_circular = TRUE,
  genes = NULL,
  gene_focus = c("fCNA", "circular"),
  palette = c("grey", "#0066CC", "#CC0033"),
  class_col = "class",
  ending_time = NULL,
  ...
)
}
\arguments{
\item{fCNA}{a \link{fCNA} object.}

\item{surv_data}{survival data, eithor a 3-column \code{data.frame} to store
sample, time and status, or a length-2 string to specify the colnames
representing time and status in \code{fCNA$sample_summary}.
\itemize{
\item sample must be identical to sample ID in \code{fCNA}.
\item time must be numeric.
\item status must be 0 or 1.
}}

\item{merge_circular}{if \code{TRUE}, merge 'possibly_circular' into 'circular' type.}

\item{genes}{a list of genes. Must be subsets of \code{fCNA$data$gene_id}.}

\item{gene_focus}{focal amplication type you focus on.
Can be 'fCNA' or 'circular'. If 'fCNA' selected,
noncircular and circular genes are included to classify samples.}

\item{palette}{plot color palette.}

\item{class_col}{column name in \code{sample_summary} field for classification.
If you set to other column (you want to run survival analysis with custom column),
parameters like \code{merge_circular}, \code{genes}, \code{gene_focus}
etc. will be omitted.}

\item{ending_time}{survival analysis ending time. If a numeric ending
is typed, all survival data longer than the ending time will be rewritten.}

\item{...}{other parameters passing to \code{survminer::ggsurvplot}.}
}
\value{
a plot.
}
\description{
See \code{\link[=gcap.plotProfile]{gcap.plotProfile()}} for examples.
}
\seealso{
\link{gcap.plotProfile} for plot landscape of fCNA, \link{fCNA} for building object.
}
