% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{gcap.plotProfile}
\alias{gcap.plotProfile}
\title{Draw Oncoprint for fCNA Profile Visualization}
\usage{
gcap.plotProfile(
  fCNA,
  genes = NULL,
  samples = NULL,
  merge_circular = FALSE,
  show_column_names = TRUE,
  remove_empty_columns = FALSE,
  remove_empty_rows = TRUE,
  heatmap_legend_param = list(title = "fCNA"),
  col = c(noncircular = "blue", circular = "red", possibly_circular = "magenta"),
  ...
)
}
\arguments{
\item{fCNA}{a \link{fCNA} object.}

\item{genes}{a list of genes. Must be subsets of \code{fCNA$data$gene_id}.}

\item{samples}{a vector to subset samples.}

\item{merge_circular}{if \code{TRUE}, merge 'possibly_circular' into 'circular' type.}

\item{show_column_names}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{remove_empty_columns}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{remove_empty_rows}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{heatmap_legend_param}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{col}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{...}{Other parameters passing to \code{ComplexHeatmap::oncoPrint}.}
}
\value{
a \code{Heatmap} object, you can go further modify it.
}
\description{
See \href{https://jokergoo.github.io/ComplexHeatmap-reference/book/oncoprint.html}{oncoprint}
and \href{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#size-of-the-heatmap}{heatmap}
to modify the result plot in depth.
}
\examples{
\donttest{
if (require("ComplexHeatmap") && require("IDConverter")) {
  data("ascn")
  data <- ascn

  # Create fake data
  set.seed(1234)
  data$sample <- sample(LETTERS[1:10], nrow(data), replace = TRUE)
  rv <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")

  gcap.plotProfile(rv, samples = c("B", "A", "C"))

  rv$convertGeneID()
  ht <- gcap.plotProfile(rv,
    samples = c("B", "A", "C", "D", "F"),
    genes = unique(rv$data$gene_id)[1:10],
    top_annotation = ComplexHeatmap::HeatmapAnnotation(
      cbar = ComplexHeatmap::anno_oncoprint_barplot(),
      foo1 = c("g1", "g1", "g2", "g2", "g3"),
      bar1 = ComplexHeatmap::anno_points(1:5)
    )
  )
  ht
  ComplexHeatmap::draw(ht, merge_legends = TRUE)

  #----------------
  # Plot KM curve
  #----------------
  surv_data <- data.frame(
    sample = rv$sample_summary$sample,
    time = 3000 * abs(rnorm(nrow(rv$sample_summary))),
    status = sample(c(0, 1), nrow(rv$sample_summary), replace = TRUE)
  )
  surv_data
  p <- gcap.plotKMcurve(rv, surv_data)
  p

  p2 <- gcap.plotKMcurve(rv, surv_data, genes = "MYC")
  p2
}
}
}
\seealso{
\link{fCNA} for building object.
}
