% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.R
\name{gcap.runASCAT}
\alias{gcap.runASCAT}
\title{Run ASCAT on tumor-normal pair WES data files}
\usage{
gcap.runASCAT(
  tumourseqfile,
  normalseqfile,
  tumourname,
  normalname,
  jobname = tumourname,
  outdir = getwd(),
  allelecounter_exe = "~/miniconda3/envs/cancerit/bin/alleleCounter",
  g1000allelesprefix = file.path("~/data/snp/1000G_loci_hg38",
    "1kg.phase3.v5a_GRCh38nounref_allele_index_chr"),
  g1000lociprefix = file.path("~/data/snp/1000G_loci_hg38",
    "1kg.phase3.v5a_GRCh38nounref_loci_chrstring_chr"),
  GCcontentfile = "~/data/snp/GC_correction_hg38.txt",
  replictimingfile = "~/data/snp/RT_correction_hg38.txt",
  nthreads = 22,
  minCounts = 10,
  BED_file = NA,
  probloci_file = NA,
  chrom_names = 1:22,
  gender = "XX",
  min_base_qual = 20,
  min_map_qual = 35,
  penalty = 70,
  skip_finished_ASCAT = FALSE
)
}
\arguments{
\item{tumourseqfile}{Full path to the tumour BAM file.}

\item{normalseqfile}{Full path to the normal BAM file.}

\item{tumourname}{Identifier to be used for tumour output files.}

\item{normalname}{Identifier to be used for normal output files.}

\item{jobname}{job name, typically an unique name for a tumor-normal pair.}

\item{outdir}{result output path.}

\item{allelecounter_exe}{Path to the allele counter executable.}

\item{g1000allelesprefix}{Prefix path to the 1000 Genomes alleles reference files.}

\item{g1000lociprefix}{Prefix path to the 1000 Genomes SNP reference files.}

\item{GCcontentfile}{File containing the GC content around every SNP for increasing window sizes}

\item{replictimingfile}{File containing replication timing at every SNP for various cell lines (optional)}

\item{nthreads}{The number of parallel processes for getting allele counts (optional, default=1).}

\item{minCounts}{Minimum depth required in the normal for a SNP to be considered (optional, default=10).}

\item{BED_file}{A BED file for only looking at SNPs within specific intervals (optional, default=NA).}

\item{probloci_file}{A file (chromosome <tab> position; no header) containing specific loci to ignore (optional, default=NA).}

\item{chrom_names}{A vector containing the names of chromosomes to be considered (optional, default=c(1:22,'X')).}

\item{gender}{a vector of gender for each cases ("XX" or "XY").
Default = all female ("XX"). Ignore this if you don't include sex
chromosomes.}

\item{min_base_qual}{Minimum base quality required for a read to be counted (optional, default=20).}

\item{min_map_qual}{Minimum mapping quality required for a read to be counted (optional, default=35).}

\item{penalty}{penalty of introducing an additional ASPCF breakpoint (expert parameter, don't adapt unless you know what you're doing)}

\item{skip_finished_ASCAT}{if \code{TRUE}, skipped finished ASCAT calls
to save time.}
}
\value{
Nothing. Check the \code{outdir} for results.
}
\description{
A wrapper calling ASCAT on WES data on one or more tumor-normal pairs.
Note, for multiple tumor-normal pairs, the first 5 arguments should
be a vector with same length.
}
