% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCNworkflow.R
\name{gcap.runASCNBuildflow}
\alias{gcap.runASCNBuildflow}
\title{Build data for prediction from absolute copy number data}
\usage{
gcap.runASCNBuildflow(data, genome_build = c("hg38", "hg19"), overlap = 1)
}
\arguments{
\item{data}{a \code{data.frame} with following columns. The key columns can be obtained
from common allele specific CNV calling software, e.g., ASCAT, Sequenza, FACETS.
\itemize{
\item chromosome: chromosome names starts with 'chr'.
\item start: start position of the segment.
\item end: end position of the segment.
\item total_cn: total integer copy number of the segment.
\item minor_cn: minor allele integer copy number of the segment. Set it
to \code{NA} if you don't have this data.
\item sample: sample identifier.
\item purity: tumor purity of the sample. Set to \code{1} if you don't know.
\item ploidy (optinal): ploidy value of the sample tumor genome.
\item age (optional): age of the case, use along with \code{gender}.
\item gender (optional): gender of the case, use along with \code{age}.
\item type (optional): cancer type of the case, use along with \code{age} and \code{gender}.
Please refer to \link{gcap.collapse2Genes} to see the supported cancer types.
This info is only used in 'XGB56' model. If you don't use this model, you
don't need to set it.
}}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}

\item{overlap}{the overlap percentage on gene.}
}
\value{
a \code{data.table}.
}
\description{
This is is a wrapper of \code{\link[=gcap.extractFeatures]{gcap.extractFeatures()}}
and \code{\link[=gcap.collapse2Genes]{gcap.collapse2Genes()}} to combine the feature extraction
and predict input generate procedure.
If you want to modify the result of \code{\link[=gcap.extractFeatures]{gcap.extractFeatures()}},
you should always use the two functions instead of this
wrapper.
}
\seealso{
\link{gcap.runBuildflow}
}
