% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{gcap.runPrediction}
\alias{gcap.runPrediction}
\title{Run gene-level circular prediction}
\usage{
gcap.runPrediction(data, model = "XGB32")
}
\arguments{
\item{data}{data to predict (\code{data.frame}/\code{matrix} format), from
\code{\link[=gcap.collapse2Genes]{gcap.collapse2Genes()}} in general.}

\item{model}{model name ("XGB11", "XGB32", "XGB56") or a custom model
from input. 'toy' can be used for test.}
}
\value{
a numeric vector representing prob.
}
\description{
Run gene-level circular prediction
}
\examples{
data("ec")
# Use toy model for illustration
y_pred <- gcap.runPrediction(ec, "toy")
y_pred
}
