% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{gcap.runScoring}
\alias{gcap.runScoring}
\title{Summarize prediction result into gene/sample-level}
\usage{
gcap.runScoring(data, prob_cutoff = 0.45)
}
\arguments{
\item{data}{a \code{data.table} containing result from \link{gcap.runPrediction}.
The column storing prediction result must be "prob".}

\item{prob_cutoff}{a numeric cutoff \verb{[0, 1]} to determine the positive class.}
}
\value{
a list of \code{data.table}.
}
\description{
Summarize prediction result into gene/sample-level
}
\examples{
data("ec")
ec2 <- ec
ec2$pred <- gcap.runPrediction(ec)
score <- gcap.runScoring(ec2)
score
}
