% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets_pipeline.R
\name{gcap.workflow.facets}
\alias{gcap.workflow.facets}
\title{GCAP FACETS workflow for gene-level amplicon prediction}
\usage{
gcap.workflow.facets(
  tumourseqfile,
  normalseqfile,
  jobname,
  extra_info = NULL,
  include_type = FALSE,
  genome_build = c("mm10", "hg38", "hg19"),
  model = "XGB11",
  tightness = 1L,
  gap_cn = 3L,
  overlap = 1,
  pro_cval = 100,
  only_oncogenes = FALSE,
  snp_file = "path/to/genome_build_responding.vcf.gz",
  outdir = getwd(),
  result_file_prefix = paste0("gcap_", uuid::UUIDgenerate(TRUE)),
  util_exe = system.file("extcode", "snp-pileup", package = "facets"),
  nthreads = 1,
  skip_finished_facets = TRUE,
  skip_facets_call = FALSE
)
}
\arguments{
\item{tumourseqfile}{Full path to the tumour BAM file.}

\item{normalseqfile}{Full path to the normal BAM file.}

\item{jobname}{job name, typically an unique name for a tumor-normal pair.}

\item{extra_info}{(optional) a (file containing) \code{data.frame} with 3 columns 'sample'
(must identical to the setting of parameter \code{jobname}),
'age' and 'gender'. For gender, should be 'XX' or 'XY',
also could be \code{0} for 'XX' and \code{1} for 'XY'.}

\item{include_type}{if \code{TRUE}, a fourth column named 'type'
should be included in \code{extra_info}, the supported cancer
type should be described with \href{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}{TCGA cancer type abbr.}.}

\item{genome_build}{genome build version, should be one of 'hg38', 'hg19' and 'mm10'.}

\item{model}{model name ("XGB11", "XGB32", "XGB56") or a custom model
from input. 'toy' can be used for test.}

\item{tightness}{a coefficient to times to TCGA somatic CN to set a more strict threshold
as a circular amplicon.
If the value is larger, it is more likely a fCNA assigned to \code{noncircular}
instead of \code{circular}. \strong{When it is \code{NA}, we don't use TCGA somatic CN data as reference}.}

\item{gap_cn}{a gap copy number value.
A gene with copy number above background (\code{ploidy + gap_cn} in general) would be treated as focal amplicon.
Smaller, more amplicons.}

\item{overlap}{the overlap percentage on gene.}

\item{pro_cval}{critical value for segmentation used in \code{facets::procSample()}.}

\item{only_oncogenes}{if \code{TRUE}, only known oncogenes are kept for circular prediction.}

\item{snp_file}{a file path to SNP file of genome, should be consistent with \code{genome_build} option.}

\item{outdir}{result output path.}

\item{result_file_prefix}{file name prefix (without directory path) for storing
final model prediction file in CSV format.
Default a unique file name is generated by UUID approach.}

\item{util_exe}{the path to \code{snp-pileup}.}

\item{nthreads}{The number of parallel processes for getting allele counts (optional, default=1).}

\item{skip_finished_facets}{if \code{TRUE}, skip finished FACETS runs.}

\item{skip_facets_call}{if \code{TRUE}, skip calling FACETS.
This is useful when you have done this step and just want
to run next steps.}
}
\value{
a list of invisible \code{data.table} and corresponding files saved to local machine.
}
\description{
GCAP FACETS workflow for gene-level amplicon prediction
}
\details{
For generating the \code{snp-pileup} program, reference commands given here.
You need modify corresponding path to fit your own machine.

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{cd /data3/wsx/R/x86_64-pc-linux-gnu-library/4.2/facets/extcode/
g++ -std=c++11 -I/data3/wsx/miniconda3/envs/circlemap/include snp-pileup.cpp -L/data3/wsx/miniconda3/envs/circlemap/lib -lhts -Wl,-rpath=/data3/wsx/miniconda3/envs/circlemap/lib -o snp-pileup
}\if{html}{\out{</div>}}
}
