% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqz_pipeline.R
\name{gcap.workflow.seqz}
\alias{gcap.workflow.seqz}
\title{GCAP sequenza workflow for gene-level amplicon prediction}
\usage{
gcap.workflow.seqz(
  tumourseqfile,
  normalseqfile,
  jobname,
  extra_info = NULL,
  include_type = FALSE,
  genome_build = c("mm10", "hg38", "hg19"),
  model = "XGB11",
  tightness = 1L,
  gap_cn = 3L,
  overlap = 1,
  only_oncogenes = FALSE,
  ref_file = "path/to/reference.fa",
  data_tmp_dir = "~/gcap_data",
  outdir = getwd(),
  result_file_prefix = paste0("gcap_", uuid::UUIDgenerate(TRUE)),
  util_exe = "~/miniconda3/bin/sequenza-utils",
  samtools_exe = "~/miniconda3/bin/samtools",
  tabix_exe = "~/miniconda3/bin/tabix",
  nthreads = 1,
  skip_finished_sequenza = TRUE,
  skip_sequenza_call = FALSE
)
}
\arguments{
\item{tumourseqfile}{Full path to the tumour BAM file.}

\item{normalseqfile}{Full path to the normal BAM file.}

\item{jobname}{job name, typically an unique name for a tumor-normal pair.}

\item{extra_info}{(optional) a (file containing) \code{data.frame} with 3 columns 'sample'
(must identical to the setting of parameter \code{jobname}),
'age' and 'gender'. For gender, should be 'XX' or 'XY',
also could be \code{0} for 'XX' and \code{1} for 'XY'.}

\item{include_type}{if \code{TRUE}, a fourth column named 'type'
should be included in \code{extra_info}, the supported cancer
type should be described with \href{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}{TCGA cancer type abbr.}.}

\item{genome_build}{genome build version, should be one of 'hg38', 'hg19' and 'mm10'.}

\item{model}{model name ("XGB11", "XGB32", "XGB56") or a custom model
from input. 'toy' can be used for test.}

\item{tightness}{a coefficient to times to TCGA somatic CN to set a more strict threshold
as a circular amplicon.
If the value is larger, it is more likely a fCNA assigned to \code{noncircular}
instead of \code{circular}. \strong{When it is \code{NA}, we don't use TCGA somatic CN data as reference}.}

\item{gap_cn}{a gap copy number value.
A gene with copy number above background (\code{ploidy + gap_cn} in general) would be treated as focal amplicon.
Smaller, more amplicons.}

\item{overlap}{the overlap percentage on gene.}

\item{only_oncogenes}{if \code{TRUE}, only known oncogenes are kept for circular prediction.}

\item{ref_file}{a reference genome file, should be consistent with \code{genome_build} option.}

\item{data_tmp_dir}{a directory path for storing temp data for reuse in handling multiple samples.}

\item{outdir}{result output path.}

\item{result_file_prefix}{file name prefix (without directory path) for storing
final model prediction file in CSV format.
Default a unique file name is generated by UUID approach.}

\item{util_exe}{the path to \code{sequenza-utils}.}

\item{samtools_exe}{the path to \code{samtools_exe}.}

\item{tabix_exe}{the path to \code{tabix}.}

\item{nthreads}{The number of parallel processes for getting allele counts (optional, default=1).}

\item{skip_finished_sequenza}{if \code{TRUE}, skip finished sequenza runs.}

\item{skip_sequenza_call}{if \code{TRUE}, skip calling sequenza.
This is useful when you have done this step and just want
to run next steps.}
}
\value{
a list of invisible \code{data.table} and corresponding files saved to local machine.
}
\description{
GCAP sequenza workflow for gene-level amplicon prediction
}
