% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{get_auc}
\alias{get_auc}
\title{Get AUC value}
\usage{
get_auc(y_pred, y, type = c("pr", "roc"), curve = FALSE)
}
\arguments{
\item{y_pred}{y prediction vector.}

\item{y}{y true label vector.}

\item{type}{AUC type, either 'pr' or 'roc'.}

\item{curve}{if \code{TRUE}, generate plot data, the result can be
plotted by \code{plot()}.}
}
\value{
A object.
}
\description{
Get AUC value
}
\examples{
if (require("PRROC")) {
  set.seed(2021)
  auc <- get_auc(sample(1:10, 10), c(rep(0, 5), rep(1, 5)))
  auc
}
}
