% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeDTs}
\alias{mergeDTs}
\title{Merge a list of data.table}
\usage{
mergeDTs(dt_list, by = NULL, sort = FALSE)
}
\arguments{
\item{dt_list}{a list of \code{data.table}s.}

\item{by}{which column used for merging.}

\item{sort}{should sort the result?}
}
\value{
a \code{data.table}
}
\description{
Merge a list of data.table
}
