% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncogenes}
\alias{oncogenes}
\title{Oncogene list}
\format{
A \code{data.frame}
}
\source{
Generate from \verb{data-raw/}, raw source from \url{http://ongene.bioinfo-minzhao.org/}
}
\description{
Oncogene list
}
\examples{
data("oncogenes")
}
