% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overlaps}
\alias{overlaps}
\title{Get overlaps of two genomic regions}
\usage{
overlaps(x, y)
}
\arguments{
\item{x, y}{a genemic region with data.frame format, the first 3 columns
should representing chromosome, start and end position.}
}
\value{
a \code{data.table}
}
\description{
Get overlaps of two genomic regions
}
