% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse2Genes.R
\name{gcap.collapse2Genes}
\alias{gcap.collapse2Genes}
\title{Generate unified gene-level feature data}
\usage{
gcap.collapse2Genes(
  fts,
  extra_info,
  include_type = FALSE,
  genome_build = c("hg38", "hg19")
)
}
\arguments{
\item{fts}{(modified) result from \code{\link[=gcap.extractFeatures]{gcap.extractFeatures()}}}

\item{extra_info}{a \code{data.frame} with 3 columns 'sample',
'age' and 'gender', for including cancer type, check
parameter \code{include_type}. For gender, should be 'XX' or 'XY',
also could be \code{0} for 'XX' and \code{1} for 'XY'.}

\item{include_type}{if \code{TRUE}, a fourth column named 'type'
should be included in \code{extra_info}, the supported cancer
type includes \code{c("BLCA", "BRCA", "CESC", "COAD", "ESCA", "GBM", "HNSC", "KICH", "KIRP", "LGG", "LIHC", "LUAD", "LUSC", "OV", "PRAD", "SARC", "SKCM", "STAD", "UCEC", "UVM")}.}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}
}
\value{
a \code{data.table}.
}
\description{
Generate unified gene-level feature data
}
