% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{gcap.plotProfile}
\alias{gcap.plotProfile}
\title{Draw Oncoprint for fCNA Profile Visualization}
\usage{
gcap.plotProfile(
  fCNA,
  genes = NULL,
  samples = NULL,
  top_n = NULL,
  top_n_by = c("circular", "Total", "noncircular"),
  only_circular = FALSE,
  merge_circular = FALSE,
  show_column_names = TRUE,
  remove_empty_columns = FALSE,
  remove_empty_rows = TRUE,
  heatmap_legend_param = list(title = "fCNA"),
  col = c(noncircular = "#0066CC", circular = "#CC0033", possibly_circular = "#FFCCCC"),
  ...
)
}
\arguments{
\item{fCNA}{a \link{fCNA} object.}

\item{genes}{a list of genes. Must be subsets of \code{fCNA$data$gene_id}.}

\item{samples}{a vector to subset samples.}

\item{top_n}{top N genes to show.}

\item{top_n_by}{how to order the genes in \code{gene_summary} to get top N genes.}

\item{only_circular}{only show circular amplicon.}

\item{merge_circular}{if \code{TRUE}, merge 'possibly_circular' into 'circular' type.}

\item{show_column_names}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{remove_empty_columns}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{remove_empty_rows}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{heatmap_legend_param}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{col}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{...}{Other parameters passing to \code{ComplexHeatmap::oncoPrint}.}
}
\value{
a \code{Heatmap} object, you can go further modify it.
}
\description{
See \href{https://jokergoo.github.io/ComplexHeatmap-reference/book/oncoprint.html}{oncoprint}
and \href{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#size-of-the-heatmap}{heatmap}
to modify the result plot in depth.
}
\examples{
\donttest{
if (require("ComplexHeatmap") && require("IDConverter")) {
  data("ascn")
  data <- ascn

  # Create fake data
  set.seed(1234)
  data$sample <- sample(LETTERS[1:10], nrow(data), replace = TRUE)
  rv <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")

  gcap.plotProfile(rv, samples = c("B", "A", "C"))

  rv$convertGeneID()
  ht <- gcap.plotProfile(rv,
    samples = c("B", "A", "C", "D", "F"),
    genes = unique(rv$data$gene_id)[1:10],
    top_annotation = ComplexHeatmap::HeatmapAnnotation(
      cbar = ComplexHeatmap::anno_oncoprint_barplot(),
      foo1 = c("g1", "g1", "g2", "g2", "g3"),
      bar1 = ComplexHeatmap::anno_points(1:5)
    )
  )
  ht
  ComplexHeatmap::draw(ht, merge_legends = TRUE)

  #----------------
  # Plot KM curve
  #----------------
  surv_data <- data.frame(
    sample = rv$sample_summary$sample,
    time = 3000 * abs(rnorm(nrow(rv$sample_summary))),
    status = sample(c(0, 1), nrow(rv$sample_summary), replace = TRUE)
  )
  surv_data
  p <- gcap.plotKMcurve(rv, surv_data)
  p

  p2 <- gcap.plotKMcurve(rv, surv_data, genes = "MYC")
  p2

  # ---------------
  # Plot forest
  # ---------------
  p3 <- gcap.plotForest(rv, surv_data)

  # Fake some data for gene analysis
  rv$data[amplicon_type == "circular"]$gene_id[1:10] <- "MYC"
  rv$data[amplicon_type == "circular"]$gene_id[11:20] <- "EGFR"
  rv$data[amplicon_type == "circular"]$gene_id[31:35] <- "GSX2"
  p4 <- gcap.plotForest(rv, surv_data,
    x = c("TP53", "MYC", "ABC", "EGFR", "GSX2"), x_is_gene = TRUE,
    ref_line = 1, xlim = c(0, 10)
  )

  # Plot on gene clusters
  gcap.plotForest(rv, surv_data,
    x = data.frame(
      cluster = paste0("cluster", c(1, 3, 2, 1, 3)),
      gene_id = c("TP53", "MYC", "ABC", "EGFR", "GSX2")
    ), x_is_gene = TRUE,
    ref_line = 1, xlim = c(0, 10), optimize_model = TRUE
  ) -> zz
  # zz$plot

  # -----------------
  # Plot distribution
  # -----------------
  gcap.plotDistribution(rv)
  p5 <- gcap.plotDistribution(rv,
    x = c("MYC", "EGFR", "CDK4", "AKT3"),
    width = 0.5, x_size = 5, fill = FALSE
  )
  p5
  rv$sample_summary[, ploidy_class := ifelse(ploidy > 2.5, "2+", "2")]
  p6 <- gcap.plotDistribution(rv, x = "ploidy_class", by = "sample")
  p6
}
}
}
\seealso{
\link{fCNA} for building object.
}
