% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildflow.R
\name{gcap.runBuildflow}
\alias{gcap.runBuildflow}
\title{Build data for prediction from ASCAT result files}
\usage{
gcap.runBuildflow(
  ascat_files,
  extra_info,
  include_type = FALSE,
  genome_build = c("hg38", "hg19", "mm10"),
  overlap = 1
)
}
\arguments{
\item{ascat_files}{a list of file path.
Typically the result of \code{\link[=gcap.runASCAT]{gcap.runASCAT()}}}

\item{extra_info}{(optional) a \code{data.frame} with 3 columns 'sample',
'age' and 'gender', for including cancer type, check
parameter \code{include_type}. For gender, should be 'XX' or 'XY',
also could be \code{0} for 'XX' and \code{1} for 'XY'.}

\item{include_type}{if \code{TRUE}, a fourth column named 'type'
should be included in \code{extra_info}, the supported cancer
type should be described with \href{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}{TCGA cancer type abbr.}.}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}

\item{overlap}{the overlap percentage on gene.}
}
\value{
a \code{data.table}.
}
\description{
This is is a wrapper of \code{\link[=gcap.extractFeatures]{gcap.extractFeatures()}}
and \code{\link[=gcap.collapse2Genes]{gcap.collapse2Genes()}} to combine the feature extraction
and predict input generate procedure.
If you want to modify the result of \code{\link[=gcap.extractFeatures]{gcap.extractFeatures()}},
you should always use the two functions instead of this
wrapper.
}
