% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{gcap.runPrediction}
\alias{gcap.runPrediction}
\title{Run gene-level amplicon prediction}
\usage{
gcap.runPrediction(
  data,
  feature = c("without_type", "with_type"),
  target = c("circle", "nonLinear"),
  use_toy = FALSE,
  custom_model = NULL
)
}
\arguments{
\item{data}{data to predict (\code{data.frame}/\code{matrix} format), from
\code{\link[=gcap.collapse2Genes]{gcap.collapse2Genes()}} in general.}

\item{feature}{one of 'with_type' and 'without_type' to select
model whether includes cancer type or not.}

\item{target}{one of 'circle' and 'nonLinear' to select model
if predict circle amplicon or non-linear amplicon.}

\item{use_toy}{if \code{TRUE}, use a toy model for showing example,
not for practical use.}

\item{custom_model}{a outer custom xgboost model for prediction.}
}
\value{
a numeric vector representing prob.
}
\description{
Run gene-level amplicon prediction
}
\examples{
data("ec")
y_pred <- gcap.runPrediction(ec)
y_pred
}
