% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{gcap.runPrediction}
\alias{gcap.runPrediction}
\title{Run gene-level amplicon prediction}
\usage{
gcap.runPrediction(
  data,
  model = "XGB32",
  target = c("circle", "nonLinear"),
  use_best_ntreelimit = TRUE
)
}
\arguments{
\item{data}{data to predict (\code{data.frame}/\code{matrix} format), from
\code{\link[=gcap.collapse2Genes]{gcap.collapse2Genes()}} in general.}

\item{model}{model name ("XGB11", "XGB32", "XGB54") or a custom model
from input. 'toy' can be used for test.}

\item{target}{one of 'circle' and 'nonLinear' to select model
if predict circle amplicon or non-linear amplicon.}

\item{use_best_ntreelimit}{if \code{TRUE}, use default \code{best_ntreelimit}
proposed by XGBOOST, otherwise we determine a iteration number (
i.e. tree number) with more careful processing to avoid over-fitting.}
}
\value{
a numeric vector representing prob.
}
\description{
Run gene-level amplicon prediction
}
\examples{
data("ec")
# Use toy model for illustration
y_pred <- gcap.runPrediction(ec, "toy")
y_pred
}
