% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{gcap.runScoring}
\alias{gcap.runScoring}
\title{Summarize prediction result into gene/sample-level}
\usage{
gcap.runScoring(data, genome_build = "hg38")
}
\arguments{
\item{data}{a \code{data.table} containing result from \link{gcap.runPrediction}.}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}
}
\value{
a list of \code{data.table}.
}
\description{
Summarize prediction result into gene/sample-level
}
\examples{
data("ec")
ec2 <- ec
ec2$prob <- gcap.runPrediction(ec)
score <- gcap.runScoring(ec2)
score
}
