% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{gcap.runScoring}
\alias{gcap.runScoring}
\title{Quantify sample-level ecDNA measueres}
\usage{
gcap.runScoring(data, cutoff = 0.9, genome_build = c("hg38", "hg19"))
}
\arguments{
\item{data}{a \code{data.table} containing result from \link{gcap.runPrediction}.
The column storing prediction result must start with \code{pred}.}

\item{cutoff}{a cutoff for converting prob into \code{0/1} value.}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}
}
\value{
a \code{data.table}.
\itemize{
\item \verb{*_load} for number of gene affected.
\item \verb{*_burden} for size of genome (per Mb) affected.
\item \verb{*_clusters} for number of region affected by clustering
distance with hard cutoff \code{1e6} (i.e., 1Mb). For genes from different
chromosomes, the distance set to \code{1e7} (i.e., 10Mb).
}
}
\description{
Quantify sample-level ecDNA measueres
}
\examples{
data("ec")
ec2 <- ec
ec2$pred <- gcap.runPrediction(ec)
score <- gcap.runScoring(ec2)
score
}
