% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{gcap.runScoring}
\alias{gcap.runScoring}
\title{Summarize prediction result into gene/sample-level}
\usage{
gcap.runScoring(
  data,
  genome_build = "hg38",
  min_prob = 0.6,
  tightness = 1L,
  gap_cn = 3L,
  only_oncogenes = FALSE
)
}
\arguments{
\item{data}{a \code{data.table} containing result from \link{gcap.runPrediction}.}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}

\item{min_prob}{the minimal aggregated (in cytoband level) probability to determine a circular amplicon.
The default value is for the balance of recall and precision. \strong{We highly recomment set it to
0.95 or larger if you want to detect solid positive cases (for experimental validation etc.)
instead of subtyping cases}.}

\item{tightness}{a coefficient to times to TCGA somatic CN to set a more strict threshold
as a circular amplicon.
If the value is larger, it is more likely a fCNA assigned to \code{noncircular}
instead of \code{circular}. \strong{When it is \code{NA}, we don't use TCGA somatic CN data as reference}.}

\item{gap_cn}{a gap copy number value, default is \code{4L} refer to Kim 2020 Nat.Gen.
A gene with copy number above \code{ploidy + gap_cn} would be treated as focal amplicon.
Smaller, more amplicons.}

\item{only_oncogenes}{if \code{TRUE}, only known oncogenes are kept for fCNA analysis
and circular prediction.}
}
\value{
a list of \code{data.table}.
}
\description{
Summarize prediction result into gene/sample-level
}
\examples{
data("ec")
ec2 <- ec
ec2$prob <- gcap.runPrediction(ec)
score <- gcap.runScoring(ec2)
score
}
