% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{gcap.runScoring}
\alias{gcap.runScoring}
\title{Summarize prediction result into gene/sample-level}
\usage{
gcap.runScoring(
  data,
  genome_build = "hg38",
  min_prob = 0.7,
  tightness = 1L,
  gap_cn = 4L
)
}
\arguments{
\item{data}{a \code{data.table} containing result from \link{gcap.runPrediction}.}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}

\item{min_prob}{the minimal aggregated (in cytoband level) probability to determine a circular amplicon.}

\item{tightness}{a coefficient to times to TCGA somatic CN to set a more strict threshold
as a circular amplicon.
If the value is larger, it is more likely a fCNA assigned to \code{noncircular}
instead of \code{circular}. \strong{When it is \code{NA}, we don't use TCGA somatic CN data as reference}.}

\item{gap_cn}{a gap copy number value, default is \code{4L} refer to Kim 2020 Nat.Gen.
A gene with copy number above \code{ploidy + gap_cn} would be treated as focal amplicon.
Smaller, more amplicons.}
}
\value{
a list of \code{data.table}.
}
\description{
Summarize prediction result into gene/sample-level
}
\examples{
data("ec")
ec2 <- ec
ec2$prob <- gcap.runPrediction(ec)
score <- gcap.runScoring(ec2)
score
}
