% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{gcap.runScoring}
\alias{gcap.runScoring}
\title{Summarize prediction result into gene/sample-level}
\usage{
gcap.runScoring(
  data,
  genome_build = "hg38",
  min_n = 1L,
  tightness = 1L,
  gap_cn = 4L,
  circ_prob = 0.75
)
}
\arguments{
\item{data}{a \code{data.table} containing result from \link{gcap.runPrediction}.}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}

\item{min_n}{a minimal cytoband number (default is \code{1}) to determine
sample class. e.g., sample with at least 1 cytoband harboring circular
genes would be labelled as "circular".}

\item{tightness}{If the value is larger, it is more likely a fCNA assigned to \code{noncircular}
instead of \code{circular}. Range from 1 to Inf.}

\item{gap_cn}{a gap copy number value, default is \code{4L} refer to Kim 2020 Nat.Gen.
A gene with copy number above \code{ploidy + gap_cn} would be treated as amplicon.
Smaller, more amplicons.}

\item{circ_prob}{a prob cutoff value to treat a gene as located at circular amplicon.}
}
\value{
a list of \code{data.table}.
}
\description{
Summarize prediction result into gene/sample-level
}
\examples{
data("ec")
ec2 <- ec
ec2$prob <- gcap.runPrediction(ec)
score <- gcap.runScoring(ec2)
score
}
