% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{gcap.workflow}
\alias{gcap.workflow}
\title{GCAP workflow for gene-level amplicon prediction}
\usage{
gcap.workflow(
  tumourseqfile,
  normalseqfile,
  tumourname,
  normalname,
  jobname = tumourname,
  extra_info,
  include_type = FALSE,
  genome_build = c("hg38", "hg19"),
  feature = c("with_type", "without_type"),
  target = c("circle", "nonLinear"),
  use_toy = FALSE,
  outdir = getwd(),
  result_file = paste0("gcap_prediction_", uuid::UUIDgenerate(TRUE), ".csv"),
  allelecounter_exe = "~/miniconda3/envs/cancerit/bin/alleleCounter",
  g1000allelesprefix = file.path("~/data/snp/1000G_loci_hg38",
    "1kg.phase3.v5a_GRCh38nounref_allele_index_chr"),
  g1000lociprefix = file.path("~/data/snp/1000G_loci_hg38",
    "1kg.phase3.v5a_GRCh38nounref_loci_chrstring_chr"),
  GCcontentfile = "~/data/snp/GC_correction_hg38.txt",
  replictimingfile = "~/data/snp/RT_correction_hg38.txt",
  nthreads = 22,
  minCounts = 10,
  BED_file = NA,
  probloci_file = NA,
  chrom_names = 1:22,
  min_base_qual = 20,
  min_map_qual = 35,
  penalty = 70,
  skip_finished_ASCAT = FALSE
)
}
\arguments{
\item{tumourseqfile}{Full path to the tumour BAM file.}

\item{normalseqfile}{Full path to the normal BAM file.}

\item{tumourname}{Identifier to be used for tumour output files.}

\item{normalname}{Identifier to be used for normal output files.}

\item{jobname}{job name, typically an unique name for a tumor-normal pair.}

\item{extra_info}{a (file containing) \code{data.frame} with 3 columns 'sample',
'age' and 'gender', for including cancer type, check
parameter \code{include_type}. For gender, should be 'XX' or 'XY',
also could be \code{0} for 'XX' and \code{1} for 'XY'.}

\item{include_type}{if \code{TRUE}, a fourth column named 'type'
should be included in \code{extra_info}, the supported cancer
type includes \code{c("BLCA", "BRCA", "CESC", "COAD", "ESCA", "GBM", "HNSC", "KICH", "KIRP", "LGG", "LIHC", "LUAD", "LUSC", "OV", "PRAD", "SARC", "SKCM", "STAD", "UCEC", "UVM")}.}

\item{genome_build}{genome build version, should be
one of 'hg38', 'hg19'.}

\item{feature}{'with_type' and 'without_type' to select model whether
includes cancer type or not. \strong{Can be one or both}.}

\item{target}{'circle' and 'nonLinear' to select model if predict circle
amplicon or non-linear amplicon. \strong{Can be one or both}.}

\item{use_toy}{if \code{TRUE}, use a toy model for showing example,
not for practical use.}

\item{outdir}{result output path.}

\item{result_file}{file name (without directory path) for storing
final model prediction file in CSV format.
Default a unique file name is generated by UUID approach.}

\item{allelecounter_exe}{Path to the allele counter executable.}

\item{g1000allelesprefix}{Prefix path to the 1000 Genomes alleles reference files.}

\item{g1000lociprefix}{Prefix path to the 1000 Genomes SNP reference files.}

\item{GCcontentfile}{File containing the GC content around every SNP for increasing window sizes}

\item{replictimingfile}{File containing replication timing at every SNP for various cell lines (optional)}

\item{nthreads}{The number of parallel processes for getting allele counts (optional, default=1).}

\item{minCounts}{Minimum depth required in the normal for a SNP to be considered (optional, default=10).}

\item{BED_file}{A BED file for only looking at SNPs within specific intervals (optional, default=NA).}

\item{probloci_file}{A file (chromosome <tab> position; no header) containing specific loci to ignore (optional, default=NA).}

\item{chrom_names}{A vector containing the names of chromosomes to be considered (optional, default=c(1:22,'X')).}

\item{min_base_qual}{Minimum base quality required for a read to be counted (optional, default=20).}

\item{min_map_qual}{Minimum mapping quality required for a read to be counted (optional, default=35).}

\item{penalty}{penalty of introducing an additional ASPCF breakpoint (expert parameter, don't adapt unless you know what you're doing)}

\item{skip_finished_ASCAT}{if \code{TRUE}, skipped finished ASCAT calls
to save time.}
}
\value{
a \code{data.table} and a file saved to local machine.
}
\description{
GCAP workflow for gene-level amplicon prediction
}
