% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_default_factor}
\alias{set_default_factor}
\title{Set default factor level for fCNA class}
\usage{
set_default_factor(class, ref_level = "nofocal")
}
\arguments{
\item{class}{a vector of fCNA class.}

\item{ref_level}{the reference level of factor.}
}
\value{
a vector.
}
\description{
Set default factor level for fCNA class
}
\examples{
set_default_factor(c("nofocal", "noncircular", "circular", "nofocal", "possibly_circular"))
}
